Rename hard-code-245 -> static-nyse-245
