+++ .changeset/lemon-foxes-lay.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/market-status-adapter': patch
+---
+
+Rename hard-code-245 to static-nyse-245
