+++ packages/composites/market-status/src/source/static-nyse-245.ts
